+++ pixelmon_neoforge21/src/main/java/com/vecoo/movelearner/api/currency/CurrencyProvider.java
@@ -1,0 +1,12 @@
+package com.vecoo.movelearner.api.currency;
+
+import com.pixelmonmod.pixelmon.api.pokemon.Pokemon;
+import com.pixelmonmod.pixelmon.battles.attacks.ImmutableAttack;
+import net.minecraft.server.level.ServerPlayer;
+import org.jetbrains.annotations.NotNull;
+
+public interface CurrencyProvider {
+    boolean buy(@NotNull ServerPlayer player, @NotNull Pokemon pokemon, @NotNull ImmutableAttack move, int price);
+
+    void successfulBuyMessage(@NotNull ServerPlayer player, @NotNull Pokemon pokemon, @NotNull ImmutableAttack move, int price);
+}