+++ pixelmon_neoforge21/src/main/java/com/vecoo/movelearner/api/events/LearnEvent.java
@@ -2,39 +2,27 @@ package com.vecoo.movelearner.api.events;
 
 import com.pixelmonmod.pixelmon.api.pokemon.Pokemon;
 import com.pixelmonmod.pixelmon.battles.attacks.ImmutableAttack;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.bus.api.Event;
 import net.neoforged.bus.api.ICancellableEvent;
 import org.jetbrains.annotations.NotNull;
 
+@Getter
+@AllArgsConstructor
 public class LearnEvent extends Event implements ICancellableEvent {
+    @NotNull
     private final ServerPlayer player;
+    @NotNull
     private final Pokemon pokemon;
+    @NotNull
     private final ImmutableAttack move;
 
-    public LearnEvent(@NotNull ServerPlayer player, @NotNull Pokemon pokemon, @NotNull ImmutableAttack move) {
-        this.player = player;
-        this.pokemon = pokemon;
-        this.move = move;
-    }
-
-    @NotNull
-    public ServerPlayer getPlayer() {
-        return this.player;
-    }
-
-    @NotNull
-    public Pokemon getPokemon() {
-        return this.pokemon;
-    }
-
-    @NotNull
-    public ImmutableAttack getMove() {
-        return this.move;
-    }
-
+    @Getter
     public static class BuyItem extends LearnEvent implements ICancellableEvent {
+        @NotNull
         private final ItemStack itemStack;
         private final int amount;
 
@@ -44,27 +32,15 @@ public class LearnEvent extends Event implements ICancellableEvent {
             this.itemStack = itemStack;
             this.amount = amount;
         }
-
-        @NotNull
-        public ItemStack getItemStack() {
-            return this.itemStack;
-        }
-
-        public int getAmount() {
-            return this.amount;
-        }
     }
 
+    @Getter
     public static class BuyCurrency extends LearnEvent implements ICancellableEvent {
         private final int price;
 
         public BuyCurrency(@NotNull ServerPlayer player, @NotNull Pokemon pokemon, @NotNull ImmutableAttack move, int price) {
             super(player, pokemon, move);
             this.price = price;
-        }
-
-        public int getPrice() {
-            return this.price;
         }
     }
 }