+++ pixelmon_neoforge21/src/main/java/com/vecoo/movelearner/config/ServerConfig.java
@@ -1,13 +1,12 @@
 package com.vecoo.movelearner.config;
 
-import com.pixelmonmod.pixelmon.api.config.api.data.ConfigPath;
-import com.pixelmonmod.pixelmon.api.config.api.yaml.AbstractYamlConfig;
-import info.pixelmon.repack.org.spongepowered.objectmapping.ConfigSerializable;
+import com.vecoo.extralib.shade.spongepowered.configurate.objectmapping.ConfigSerializable;
+import lombok.Getter;
 
-@ConfigPath("config/MoveLearner/config.yml")
+@Getter
 @ConfigSerializable
-public class ServerConfig extends AbstractYamlConfig {
-    private boolean useCurrency = false;
+public class ServerConfig {
+    private String currencyType = "item";
     private String itemPriceMove = "pixelmon:rare_candy";
     private boolean itemStrongTags = true;
     private boolean hideAlreadyMove = false;
@@ -19,52 +18,4 @@ public class ServerConfig extends AbstractYamlConfig {
     private int tutorMovePrice = 16;
     private int transferMovePrice = 16;
     private int eggMovePrice = 32;
-
-    public boolean isUseCurrency() {
-        return this.useCurrency;
-    }
-
-    public String getItemPriceMove() {
-        return this.itemPriceMove;
-    }
-
-    public boolean isItemStrongTags() {
-        return this.itemStrongTags;
-    }
-
-    public boolean isHideAlreadyMove() {
-        return this.hideAlreadyMove;
-    }
-
-    public boolean isHmMove() {
-        return this.hmMove;
-    }
-
-    public boolean isEggMove() {
-        return this.eggMove;
-    }
-
-    public int getLevelMovePrice() {
-        return this.levelMovePrice;
-    }
-
-    public int getTmTrMovePrice() {
-        return this.tmTrMovePrice;
-    }
-
-    public int getHmMovePrice() {
-        return this.hmMovePrice;
-    }
-
-    public int getTutorMovePrice() {
-        return this.tutorMovePrice;
-    }
-
-    public int getTransferMovePrice() {
-        return this.transferMovePrice;
-    }
-
-    public int getEggMovePrice() {
-        return this.eggMovePrice;
-    }
 }