+++ pixelmon_neoforge21/src/main/java/com/vecoo/movelearner/ui/pages/SelectMovePage.java
@@ -8,15 +8,15 @@ import com.pixelmonmod.pixelmon.enums.technicalmoves.ITechnicalMove;
 import com.vecoo.extralib.chat.UtilChat;
 import com.vecoo.extralib.ui.api.gui.SimpleGui;
 import com.vecoo.movelearner.MoveLearner;
-import com.vecoo.movelearner.api.factory.MoveLearnerFactoryUI;
+import com.vecoo.movelearner.api.service.MoveLearnerServiceUI;
 import com.vecoo.movelearner.config.GuiConfig;
 import com.vecoo.movelearner.config.ServerConfig;
 import com.vecoo.movelearner.ui.Buttons;
 import com.vecoo.movelearner.ui.settings.MoveFilter;
+import lombok.Getter;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.inventory.MenuType;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -25,13 +25,17 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
+@Getter
 public class SelectMovePage extends SimpleGui {
-    private final ServerConfig CONFIG = MoveLearner.getInstance().getConfig();
+    private final ServerConfig CONFIG = MoveLearner.getInstance().getServerConfig();
     private final GuiConfig GUI_CONFIG = MoveLearner.getInstance().getGuiConfig();
 
     private final Pokemon pokemon;
+    @NotNull
     private final MoveFilter filter;
+    @NotNull
     private final String search;
+    @NotNull
     private final List<ImmutableAttack> moves;
     private final int page;
     private final int totalPages;
@@ -61,48 +65,20 @@ public class SelectMovePage extends SimpleGui {
         addNextPageButton();
     }
 
-    @Nullable
-    public Pokemon getPokemon() {
-        return this.pokemon;
-    }
-
-    @NotNull
-    public MoveFilter getFilter() {
-        return this.filter;
-    }
-
-    @NotNull
-    public String getSearch() {
-        return this.search;
-    }
-
-    @NotNull
-    public List<ImmutableAttack> getMoves() {
-        return this.moves;
-    }
-
-    public int getPage() {
-        return this.page;
-    }
-
-    public int getTotalPages() {
-        return this.totalPages;
-    }
-
     private void openPage() {
-        MoveLearnerFactoryUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, this.filter, this.search, this.page));
+        MoveLearnerServiceUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, this.filter, this.search, this.page));
     }
 
     private void openPage(int newPage) {
-        MoveLearnerFactoryUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, this.filter, this.search, newPage));
+        MoveLearnerServiceUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, this.filter, this.search, newPage));
     }
 
     private void openPage(@NotNull MoveFilter filter) {
-        MoveLearnerFactoryUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, filter, "", 1));
+        MoveLearnerServiceUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, filter, "", 1));
     }
 
     private void openPage(@NotNull String search) {
-        MoveLearnerFactoryUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, MoveFilter.ALL, search, 1));
+        MoveLearnerServiceUI.openPage(player, this.pokemon, new SelectMovePage(player, this.pokemon, MoveFilter.ALL, search, 1));
     }
 
     @NotNull
@@ -125,7 +101,7 @@ public class SelectMovePage extends SimpleGui {
 
         for (ImmutableAttack move : moves) {
             setSlot(slot++, Buttons.getMoveButton(this.pokemon, move, player)
-                    .setCallback(() -> MoveLearnerFactoryUI.openPage(player, this.pokemon, new AcceptPage(player, move, this))));
+                    .setCallback(() -> MoveLearnerServiceUI.openPage(player, this.pokemon, new AcceptPage(player, move, this))));
         }
     }